Moved validateRecipeName into recipe-book.ts module and deleted form-validation module
