+++ tests/unit-tests/recipe-book.test.ts
@@ -3,6 +3,7 @@ import {
   cloneRecipe,
   deleteRecipe,
   updateRecipe,
+  validateRecipeName,
 } from "../../src/core/recipe-book";
 
 describe("cloneRecipe", () => {
@@ -151,5 +152,69 @@ describe("updateRecipe", () => {
     expect(newRecipeBook).toStrictEqual({
       "Quinoa Broccoli Casserole": newRecipe,
     });
+  });
+});
+
+describe("validateRecipeName", () => {
+  it("returns false for the empty string", () => {
+    expect(validateRecipeName("old recipe name", "", {})).toStrictEqual(
+      "Recipe name cannot be empty"
+    );
+  });
+
+  it("returns false for the strings longer than 100 characters", () => {
+    expect(
+      validateRecipeName(
+        "old recipe name",
+        "This is a super long recipe name that is one hundred and one characters long; this is not readable!!!",
+        {}
+      )
+    ).toStrictEqual("Recipe name cannot be longer than 100 characters");
+  });
+
+  it("returns false if this recipe name already exists in the recipe book", () => {
+    const recipeBook: RecipeBook = {
+      "Quinoa Broccoli Casserole": {
+        yield: { amount: 10, units: "servings" },
+        ingredients: [],
+        notes: "",
+      },
+    };
+
+    expect(
+      validateRecipeName("", "Quinoa Broccoli Casserole", recipeBook)
+    ).toStrictEqual("A recipe with this name already exists");
+  });
+
+  it("returns true if this is a brand new recipe name", () => {
+    const recipeBook: RecipeBook = {
+      "Blueberry pie": {
+        yield: { amount: 10, units: "servings" },
+        ingredients: [],
+        notes: "",
+      },
+    };
+
+    expect(
+      validateRecipeName("", "Quinoa Broccoli Casserole", recipeBook)
+    ).toStrictEqual("");
+  });
+
+  it("returns true if the recipe name is unchanged", () => {
+    const recipeBook: RecipeBook = {
+      "Quinoa Broccoli Casserole": {
+        yield: { amount: 10, units: "servings" },
+        ingredients: [],
+        notes: "",
+      },
+    };
+
+    expect(
+      validateRecipeName(
+        "Quinoa Broccoli Casserole",
+        "Quinoa Broccoli Casserole",
+        recipeBook
+      )
+    ).toStrictEqual("");
   });
 });
